Fixed some conversion errors that occur when the -Wc++-compat compiler flag is enabled.

+++ include/mqtt.h
@@ -283,7 +283,7 @@ uint16_t __mqtt_unpack_uint16(const uint8_t *buf);
 ssize_t __mqtt_pack_str(uint8_t *buf, const char* str);
 
 /** @brief A macro to get the MQTT string length from a c-string. */
-#define __mqtt_packed_cstrlen(x) (2 + strlen(x))
+#define __mqtt_packed_cstrlen(x) (2 + (unsigned int)strlen(x))
 
 /* RESPONSES */
